requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.16
django-autoslug==1.9.8
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==2.4.0
django-storages==1.10.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.3.3
MarkupPy==1.14
openpyxl==3.0.5
Pillow==7.2.0
pytz==2020.1
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
PyYAML==5.3.1
odfpy==1.4.1
